refactor: split tauri-free backend into linxiv-server
Router/state/sync + headless/dev-server bins move to crates/server;
the app keeps thin command wrappers. No webkit in headless builds.
