+++ src-tauri/crates/server/src/headless_admin.html
