+++ src-tauri/crates/server/src/route/editor.rs
