+++ src-tauri/crates/server/src/route/pdfs.rs
