+++ src-tauri/crates/server/src/route/projects.rs
