+++ src-tauri/crates/server/src/route/search.rs
