+++ src-tauri/crates/server/src/route/settings.rs
