+++ src-tauri/crates/server/src/route/trash.rs
