+++ src-tauri/src/lib.rs
@@ -1,14 +1,14 @@
-//! Shared in-process backend for the linXiv desktop app: the HTTP-shaped router
-//! over `linxiv-core` plus app state, the `linxiv://` protocol handler, and the
-//! CLI/MCP-integration commands. Consumed by two binaries — the Tauri app
-//! (`src/main.rs`) and the dev-only HTTP shim (`src/bin/dev_server.rs`, D32).
+//! Tauri shell of the linXiv desktop app: the invoke-command layer over
+//! `linxiv-server` (which owns the router, state, and background tasks), the
+//! `linxiv://` protocol handler, and the CLI/MCP-integration commands.
+//! Consumed by `src/main.rs`; the headless/dev-server bins live in
+//! `crates/server` and never touch this crate.
 
-pub mod full_text_worker;
+pub mod commands;
 pub mod integrations;
-pub mod p2p_config;
 pub mod protocol;
 pub mod remote_backend;
-pub mod remote_query;
-pub mod route;
-pub mod share_sync;
-pub mod state;
+
+// The shared backend, re-exported so app modules (and tests) keep their
+// `crate::route`-style paths.
+pub use linxiv_server::{full_text_worker, p2p_config, remote_query, route, share_sync, state};
